moved new test to the test_factory file
